Unity: change roomIdList to List<int>

+++ client-unity/Assets/2 - Scripts/components/RoomIdList.cs
@@ -7,12 +7,13 @@ public class RoomIdList : MonoBehaviour
     [SerializeField]
     private GameObject roomButtonPrefab;
 
-    public void SetRoomIdList(List<long> roomIdList)
+    public void SetRoomIdList(List<int> roomIdList)
     {
         roomIdList.Sort();
         Debug.Log("RoomIdList.SetRoomIdList: " + string.Join(",", roomIdList));
         gameObject.GetComponent<ListUI>().RemoveAllItems();
-        foreach (long roomId in roomIdList)
+        foreach (int roomId in roomIdList)
+
         {
             GameObject go = gameObject.GetComponent<ListUI>().AddItem(roomButtonPrefab);
             go.GetComponentInChildren<Text>().text = "Room #" + roomId;
